Move the common functions in separate files
