+++ plugins/wfp/management/commands/wfp_etl.py
@@ -1,10 +1,11 @@
-from ...models import *  # type: ignore
 from django.core.management.base import BaseCommand
-from ...tasks import ETL
+from wfp_etl_Under5 import Under5
+from wfp_etl_pbwg import PBWG
 
 
 class Command(BaseCommand):
     help = "Transform WFP collected data in a format usable for analytics"
 
     def handle(self, *args, **options):
-        ETL().run()
+        Under5.run()
+        PBWG.run()
